- I created separate files for all the controllers and renamed them, changed the references to point to the new controllers. - I created new server side method for paging, create a new loadMore method for paging in the front home controller and connected ionic infinite scroll with new logic.
