+++ server/app/controllers/user.js
@@ -1,8 +1,8 @@
 // Load the Required packages
-var User 		 			= require('../models/user');
-var Article		 		= require('../models/article.js');
-var Notification  = require('../models/notification.js');
-var mongoose 	 	  = require('mongoose');
+var User = require('../models/user');
+var Article = require('../models/article.js');
+var Notification = require('../models/notification.js');
+var mongoose = require('mongoose');
 
 // 	  TABLE OF CONTENTS   ====================================================================
 //
@@ -21,33 +21,33 @@ var mongoose 	 	  = require('mongoose');
 // ====================================================
 
 // POST
-exports.postUsers = function(req, res) {
-	var user = new User({
-		email: req.body.email,
-		username: req.body.username,
-		description: '',
-		gravatarURL: req.body.gravatarURL,
-		authID: req.body.authID,
-		counts: {
-			articles: 0,
-			follows: 0,
-			followed_by: 0
-		}
-	});
-
-	user.save(function(err) {
-		if (err)
-			res.send(err);
-
-		res.json(
+exports.postUsers = function (req, res) {
+    var user = new User({
+        email: req.body.email,
+        username: req.body.username,
+        description: '',
+        gravatarURL: req.body.gravatarURL,
+        authID: req.body.authID,
+        counts: {
+            articles: 0,
+            follows: 0,
+            followed_by: 0
+        }
+    });
+    
+    user.save(function (err) {
+        if (err)
+            res.send(err);
+        
+        res.json(
 			{
-				message: 'New User has been added!',
-				email: user.email,
-				username: user.username,
-				gravatarURL: user.gravatarURL
-			}
-		);
-	});
+            message: 'New User has been added!',
+            email: user.email,
+            username: user.username,
+            gravatarURL: user.gravatarURL
+        }
+);
+    });
 };
 
 
@@ -55,15 +55,15 @@ exports.postUsers = function(req, res) {
 //               #getUser
 // ====================================================
 
-exports.getUser = function(req, res) {
-	var id = req.params.id;
-
-	User.findById(id, function(err, users) {
-		if (err)
-			res.send(err);
-		res.json(users);
+exports.getUser = function (req, res) {
+    var id = req.params.id;
+    
+    User.findById(id, function (err, users) {
+        if (err)
+            res.send(err);
+        res.json(users);
 		// console.log(users);
-	});
+    });
 
 };
 
@@ -72,29 +72,29 @@ exports.getUser = function(req, res) {
 //               #putUser
 // ====================================================
 
-exports.putUser = function(req, res) {
-
-	var userID = req.body.userID;
-	var username = req.body.username;
-	var email = req.body.email;
-	var description = req.body.description;
-
-	User.findById(userID, function(err, user) {
-    if (err)
-      res.send(err);
-
-    user.username = username;
-		user.email = email;
-		user.description = description;
-
-
-    user.save(function(err) {
-      if (err)
-        res.send(err);
-
-      res.json(user);
+exports.putUser = function (req, res) {
+    
+    var userID = req.body.userID;
+    var username = req.body.username;
+    var email = req.body.email;
+    var description = req.body.description;
+    
+    User.findById(userID, function (err, user) {
+        if (err)
+            res.send(err);
+        
+        user.username = username;
+        user.email = email;
+        user.description = description;
+        
+        
+        user.save(function (err) {
+            if (err)
+                res.send(err);
+            
+            res.json(user);
+        });
     });
-  });
 
 };
 
@@ -103,41 +103,74 @@ exports.putUser = function(req, res) {
 // ====================================================
 
 // GET
-exports.getHomeFeed = function(req, res) {
-	console.log(req.body);
-
-	var userID = mongoose.Types.ObjectId(req.params.id);
-	var feedCount = req.body.count;
-	var minDate   = req.body.minID;
-
-	console.log("the userID is " + userID);
-	console.log("the COUNT is " + feedCount);
-
-	User
-		.findById(userID).exec(function(err, user) {
-
-					// Store user.follows in followIDlist
-					FollowIDs = [];
-
-					// Fill FollowsID array with followers
-		      for (var i = 0; i < user.follows.length; i++) {
-						FollowIDs.push(user.follows[i]);
-		      };
-
-					Article
+exports.getHomeFeed = function (req, res) {
+    console.log(req.body);
+    
+    var userID = mongoose.Types.ObjectId(req.params.id);
+    var feedCount = req.body.count;
+    var minDate = req.body.minID;
+    
+    console.log("the userID is " + userID);
+    console.log("the COUNT is " + feedCount);
+    
+    User
+		.findById(userID).exec(function (err, user) {
+        
+        // Store user.follows in followIDlist
+        FollowIDs = [];
+        
+        // Fill FollowsID array with followers
+        for (var i = 0; i < user.follows.length; i++) {
+            FollowIDs.push(user.follows[i]);
+        }        ;
+        
+        Article
 						.where('_userID').in(FollowIDs)
-						.find({ "created": {$lt: minDate}})
+						.find({ "created": { $lt: minDate } })
 						.limit(feedCount)
-						.sort({created: 'desc'})
-						.exec(function(err, articles) {
-								res.send(articles)
-						});
+						.sort({ created: 'desc' })
+						.exec(function (err, articles) {
+            res.send(articles);
+        });
 
-		});
+    });
 
 };
 
+// ====================================================
+//               #getHomeFeed with paging
+// ====================================================
+exports.getHomeFeedPaging = function (req, res) {
+    console.log(req.body);
+    
+    var userID = mongoose.Types.ObjectId(req.params.id);
+    
+    var minDate = req.body.minID;
+    var itemsPerPage = req.body.itemsPerPage;
+    var pageNumber = req.body.pageNumber;
 
+    User
+    .findById(userID)
+    .exec(function (err, user) {
+        // Store user.follows in followIDlist
+        FollowIDs = [];
+        
+        // Fill FollowsID array with followers
+        for (var i = 0; i < user.follows.length; i++) {
+            FollowIDs.push(user.follows[i]);
+        }
+        
+        Article.where('_userID').in(FollowIDs)
+			   //.find({ "created": { $lt: minDate } })
+               .skip(itemsPerPage * (pageNumber-1))
+               .limit(itemsPerPage)
+			   .sort({ created: 'desc' })
+			   .exec(function (err, articles) {
+                    return res.send(articles);
+        });
+
+    });
+};
 
 
 
@@ -146,12 +179,12 @@ exports.getHomeFeed = function(req, res) {
 //               #getUsers
 // ====================================================
 
-exports.getUsers = function(req, res) {
-	User.find(function(err, users) {
-			if (err)
-				res.send(err);
-		res.json(users);
-	});
+exports.getUsers = function (req, res) {
+    User.find(function (err, users) {
+        if (err)
+            res.send(err);
+        res.json(users);
+    });
 };
 
 
@@ -169,20 +202,20 @@ exports.getUsers = function(req, res) {
 // ====================================================
 
 // GET
-exports.getUserArticles = function(req, res) {
-
-	var userID = mongoose.Types.ObjectId(req.params.id);
-
-
-	User.findById(userID)
+exports.getUserArticles = function (req, res) {
+    
+    var userID = mongoose.Types.ObjectId(req.params.id);
+    
+    
+    User.findById(userID)
 		// If this doesnt work uncomment this part
 		// .populate('articles')
-		.populate({path: 'articles', options: { sort: { 'created': -1 } } })
+		.populate({ path: 'articles', options: { sort: { 'created': -1 } } })
 		// Need to return this most recent first
 		// .sort({created: 'desc'})
-		.exec(function(err, user) {
-    	res.send(user.articles)
-		});
+		.exec(function (err, user) {
+        res.send(user.articles)
+    });
 
 };
 
@@ -193,61 +226,61 @@ exports.getUserArticles = function(req, res) {
 // ====================================================
 
 // POST
-exports.saveForLater = function(req, res) {
-
-	var userID 		= mongoose.Types.ObjectId(req.body.userID);
-	var articleID = mongoose.Types.ObjectId(req.body.articleID);
-
-	console.log("The user id is " + userID);
-	console.log("The article id is " + articleID);
-
-	// Note to Self:for some reason this is returning null
-	User.findByIdAndUpdate(
+exports.saveForLater = function (req, res) {
+    
+    var userID = mongoose.Types.ObjectId(req.body.userID);
+    var articleID = mongoose.Types.ObjectId(req.body.articleID);
+    
+    console.log("The user id is " + userID);
+    console.log("The article id is " + articleID);
+    
+    // Note to Self:for some reason this is returning null
+    User.findByIdAndUpdate(
 			userID,
-			{$push: {"saved": articleID}},
-			{safe: true, upsert: true},
-			function(err, model) {
-					// console.log(err);
-					res.send(err)
-			}
-	);
+			{ $push: { "saved": articleID } },
+			{ safe: true, upsert: true },
+			function (err, model) {
+        // console.log(err);
+        res.send(err)
+    }
+);
 
 };
 
 
 // GET
-exports.getSaved = function(req, res) {
-
-	var userID = mongoose.Types.ObjectId(req.params.id);
-	console.log("The user id that is in question is " + userID);
-
-	User.findById(userID).populate('saved').exec(function(err, user) {
-    	res.send(user.saved)
-	});
+exports.getSaved = function (req, res) {
+    
+    var userID = mongoose.Types.ObjectId(req.params.id);
+    console.log("The user id that is in question is " + userID);
+    
+    User.findById(userID).populate('saved').exec(function (err, user) {
+        res.send(user.saved)
+    });
 
 };
 
 // DELETE
-exports.deleteSaved = function(req, res) {
-
-
-	var userID 		= mongoose.Types.ObjectId(req.body.userID);
-	var articleID = mongoose.Types.ObjectId(req.body.articleID);
-
-	console.log("The user id is " + userID);
-	console.log("The article id is " + articleID);
-
-
-	User.findByIdAndUpdate(
+exports.deleteSaved = function (req, res) {
+    
+    
+    var userID = mongoose.Types.ObjectId(req.body.userID);
+    var articleID = mongoose.Types.ObjectId(req.body.articleID);
+    
+    console.log("The user id is " + userID);
+    console.log("The article id is " + articleID);
+    
+    
+    User.findByIdAndUpdate(
 			userID,
-			{$pull: {"saved": articleID}},
-			{safe: true, upsert: true},
-			function(err, user) {
-				if (err)
-			res.send(err);
-		res.json(user);
-			}
-	);
+			{ $pull: { "saved": articleID } },
+			{ safe: true, upsert: true },
+			function (err, user) {
+        if (err)
+            res.send(err);
+        res.json(user);
+    }
+);
 
 
 };
@@ -260,101 +293,100 @@ exports.deleteSaved = function(req, res) {
 // ====================================================
 
 // POST
-exports.postFollows = function(req, res) {
-	var myID = mongoose.Types.ObjectId(req.params.id);
-	var userID = mongoose.Types.ObjectId(req.body._id);
-	// add kai's userID to pat's followers
-	console.log(myID);
-	console.log(userID);
-
-		// FEATURE TO ADD
-		// IF the user already exists, do not push them to the database
-
-		// ADDS user to my follows List
+exports.postFollows = function (req, res) {
+    var myID = mongoose.Types.ObjectId(req.params.id);
+    var userID = mongoose.Types.ObjectId(req.body._id);
+    // add kai's userID to pat's followers
+    console.log(myID);
+    console.log(userID);
+    
+    // FEATURE TO ADD
+    // IF the user already exists, do not push them to the database
+    
+    // ADDS user to my follows List
     User.findByIdAndUpdate(
         myID,
-        {$push: {"follows": userID}},
-        {safe: true, upsert: true},
-        function(err, user) {
-        	if (err)
-				res.send(err);
+        { $push: { "follows": userID } },
+        { safe: true, upsert: true },
+        function (err, user) {
+        if (err)
+            res.send(err);
 				// comment this out because in this case we want to return bob instead
 			// res.json(user);
-        }
-    );
-
-		// ADDS alices name to bobs followers array
+    }
+);
+    
+    // ADDS alices name to bobs followers array
     User.findByIdAndUpdate(
         userID,
-        {$push: {"followers": myID}},
-        {safe: true, upsert: true},
-        function(err, user) {
-        	if (err)
-					res.send(err);
-					res.json(user);
-        }
-    );
-
-
-		// Look up alice by ID and return her info
-		User.findById(myID, function(err, users) {
-			if (err)
-				res.send(err);
-			res.json(users);
-			var doer_username = users.username;
-			console.log(doer_username);
-			// function to send notification
-			postNotification(doer_username);
-		});
-
-		// send notification to bob that alice is following him
-		// build notificaiton object
-		// post notification object to bobs notification array
-
-		// NEEDS
-		// alice username
-		// alice id
-
-
-		function postNotification(doer_username) {
-
-
-				// BUILD THE NOTIFICATION OBJECT
-				var notification = new Notification({
-						doer_id: myID,
-						doer_username: doer_username,
-						articleOwner: userID,
-						type: "is now following you",
-						created: Date.now()
-				});
-
-				// PUSH NOTIFICATION TO USERS COLLECTION
-				User
+        { $push: { "followers": myID } },
+        { safe: true, upsert: true },
+        function (err, user) {
+        if (err)
+            res.send(err);
+        res.json(user);
+    }
+);
+    
+    
+    // Look up alice by ID and return her info
+    User.findById(myID, function (err, users) {
+        if (err)
+            res.send(err);
+        res.json(users);
+        var doer_username = users.username;
+        console.log(doer_username);
+        // function to send notification
+        postNotification(doer_username);
+    });
+    
+    // send notification to bob that alice is following him
+    // build notificaiton object
+    // post notification object to bobs notification array
+    
+    // NEEDS
+    // alice username
+    // alice id
+    
+    
+    function postNotification(doer_username) {
+        
+        
+        // BUILD THE NOTIFICATION OBJECT
+        var notification = new Notification({
+            doer_id: myID,
+            doer_username: doer_username,
+            articleOwner: userID,
+            type: "is now following you",
+            created: Date.now()
+        });
+        
+        // PUSH NOTIFICATION TO USERS COLLECTION
+        User
 				.findByIdAndUpdate(
 						notification.articleOwner,
-						{$push: {"notifications": notification}},
-						{safe: true, upsert: true},
-						function(err, user) {
-							if (err)
-						res.send(err);
-							res.json(notification);
-						}
-				);
-
-
-				// SAVE NOTIFICATION OBJECT
-				notification.save(function(err) {
-					if (err)
-						res.send(err);
-					res.json(
+						{ $push: { "notifications": notification } },
+						{ safe: true, upsert: true },
+						function (err, user) {
+            if (err)
+                res.send(err);
+            res.json(notification);
+        }
+);
+        
+        
+        // SAVE NOTIFICATION OBJECT
+        notification.save(function (err) {
+            if (err)
+                res.send(err);
+            res.json(
 						{
-							message: 'New Notication Has been added',
+                message: 'New Notication Has been added',
+            }
+);
+        });
 
-						}
-					);
-				});
-
-		}
+    }
 
 
 
@@ -365,49 +397,49 @@ exports.postFollows = function(req, res) {
 
 
 // GET
-exports.getFollows = function(req, res) {
-
-	var id = mongoose.Types.ObjectId(req.params.id);
-
-	User
+exports.getFollows = function (req, res) {
+    
+    var id = mongoose.Types.ObjectId(req.params.id);
+    
+    User
 		.findById(id)
 		.select('-__v -authID -email -followers -counts -articles -email')
 		.populate('follows', '-__v -authID -followers -counts -articles -follows -email')
-		.exec(function(err, user) {
-			res.send(user.follows)
-		});
+		.exec(function (err, user) {
+        res.send(user.follows)
+    });
 
 };
 
 
 
 // DELETE
-exports.deleteFollows = function(req, res) {
-
-	var myID = mongoose.Types.ObjectId(req.params.id);
-	var userID = mongoose.Types.ObjectId(req.body._id);
-
+exports.deleteFollows = function (req, res) {
+    
+    var myID = mongoose.Types.ObjectId(req.params.id);
+    var userID = mongoose.Types.ObjectId(req.body._id);
+    
     User.findByIdAndUpdate(
         myID,
-        {$pull: {"follows": userID}},
-        {safe: true, upsert: true},
-        function(err, user) {
-        	if (err)
-				res.send(err);
-			res.json(user);
-        }
-    );
-
+        { $pull: { "follows": userID } },
+        { safe: true, upsert: true },
+        function (err, user) {
+        if (err)
+            res.send(err);
+        res.json(user);
+    }
+);
+    
     User.findByIdAndUpdate(
         userID,
-        {$pull: {"followers": myID}},
-        {safe: true, upsert: true},
-        function(err, user) {
-        	if (err)
-				res.send(err);
-			res.json(user);
-        }
-    );
+        { $pull: { "followers": myID } },
+        { safe: true, upsert: true },
+        function (err, user) {
+        if (err)
+            res.send(err);
+        res.json(user);
+    }
+);
 
 };
 
@@ -418,13 +450,13 @@ exports.deleteFollows = function(req, res) {
 
 
 
-exports.getFollowers = function(req, res) {
-	var id = mongoose.Types.ObjectId(req.params.id);
-
-	console.log(id);
-	User.findById(id).populate('followers').exec(function(err, user) {
-    	res.send(user.followers)
-	});
+exports.getFollowers = function (req, res) {
+    var id = mongoose.Types.ObjectId(req.params.id);
+    
+    console.log(id);
+    User.findById(id).populate('followers').exec(function (err, user) {
+        res.send(user.followers)
+    });
 
 };
 
@@ -433,31 +465,31 @@ exports.getFollowers = function(req, res) {
 //               /users/:id/feed
 // ====================================================
 
-	// .where('userID')
-	// .in(arrayOfFollowIDs)
+// .where('userID')
+// .in(arrayOfFollowIDs)
 
 
-exports.getUserFeed = function(req, res) {
-
-	var userID = mongoose.Types.ObjectId(req.params.id);
-
-
-	User
+exports.getUserFeed = function (req, res) {
+    
+    var userID = mongoose.Types.ObjectId(req.params.id);
+    
+    
+    User
 		.findById(userID)
-		.exec(function(err, user) {
-
-			var data = user.follows;
-			// console.log(data);
-			// now the list of userIDs is in an array called data
-			// we need to pass this into the other query
-			// search those susers "recent array", or just get their most recent
-
-			Article
+		.exec(function (err, user) {
+        
+        var data = user.follows;
+        // console.log(data);
+        // now the list of userIDs is in an array called data
+        // we need to pass this into the other query
+        // search those susers "recent array", or just get their most recent
+        
+        Article
 				.where('_userID').in(data)
-				.exec(function(err, user) {
-					res.send(Article)
-			});
-	});
+				.exec(function (err, user) {
+            res.send(Article)
+        });
+    });
 
 
 
@@ -468,18 +500,18 @@ exports.getUserFeed = function(req, res) {
 //               /users/auth/:authID
 // ====================================================
 
-exports.getAuth = function(req, res) {
-
-	var id = req.params.id;
-	console.log(id);
-
-	User.findOne({
-		authID: id
-	}, function(err, users) {
-		if (err)
-			res.send(err);
-			console.log("There was an error finding a user with that auth id")
-			res.json(users);
-	});
+exports.getAuth = function (req, res) {
+    
+    var id = req.params.id;
+    console.log(id);
+    
+    User.findOne({
+        authID: id
+    }, function (err, users) {
+        if (err)
+            res.send(err);
+        console.log("There was an error finding a user with that auth id")
+        res.json(users);
+    });
 
 };
