+++ server/app/models/article.js
@@ -2,7 +2,7 @@
 // load the things we need
 var mongoose 	 = require('mongoose');
 var Schema       = mongoose.Schema;
-var deepPopulate = require('mongoose-deep-populate');
+//var deepPopulate = require('mongoose-deep-populate');
 
 var articleSchema = new Schema({
   _userID: { type: Schema.Types.ObjectId, ref: 'User' },
@@ -21,6 +21,6 @@ var articleSchema = new Schema({
   comments : [{ type: Schema.Types.ObjectId, ref: 'Comment' }]
 });
 
-articleSchema.plugin(deepPopulate);
+//articleSchema.plugin(deepPopulate);
 // create the model for users and expose it to our app
 module.exports = mongoose.model('Article', articleSchema);
