+++ server/app/models/notification.js
@@ -2,7 +2,7 @@
 // load the things we need
 var mongoose 	 = require('mongoose');
 var Schema       = mongoose.Schema;
-var deepPopulate = require('mongoose-deep-populate');
+//var deepPopulate = require('mongoose-deep-populate');
 
 
 var notificationSchema = new Schema({
@@ -17,6 +17,6 @@ var notificationSchema = new Schema({
   read: Boolean
 });
 
-notificationSchema.plugin(deepPopulate);
+//notificationSchema.plugin(deepPopulate);
 // create the model for users and expose it to our app
 module.exports = mongoose.model('Notification', notificationSchema);
