+++ sourceTabs/www/js/Controllers/homeController.js
@@ -1,0 +1,326 @@
+angular.module('starter.controllers')
+.controller('HomeController', function ($rootScope, $scope, auth, API) {
+    $scope.isLiked = true;
+    $scope.pageNumber = 1;
+    $scope.itemsPerPage = 10;
+    $scope.noMoreItemsAvailable = false;
+    $scope.articles = [];
+
+    // On before you ender the pag run this function
+    //$scope.$on('$ionicView.beforeEnter', function () {
+    // //   $scope.getInitialHomeFeed();
+    //});
+
+    // Not sure if I need this
+    $rootScope.auth = auth;
+
+    $scope.getInitialHomeFeed = function () {
+        var userID = window.localStorage.SourceID;
+        var count = 10;
+        // var minID               = date object of latest article;
+
+        var homeFeedPacket = {
+            userID: userID,
+            count: count,
+            minID: Date.now(),
+        };
+
+
+        API.getSaved(userID)
+          .success(function (data, status, headers, config) {
+              $scope.savedArticlesIds = [];
+              // This is where you put the conditional for zero states
+              for (var i = 0; i < data.length; i++) {
+                  $scope.savedArticlesIds.push(data[i]._id);
+              };
+              //  var savedArticleListIds = $scope.savedArticlesIds
+              //  console.log(savedArticleListIds);
+          })
+          .error(function (users, status, headers, config) {
+              console.log("Something went when getting list of saved articles")
+
+          });
+
+
+
+        // Stores a list of your saved articles in local storage
+        // $scope.storeSavedArticles();
+
+
+        //API.getHomeFeed(homeFeedPacket)
+        //  .success(function (data, user, status, headers, config) {
+
+        //      $scope.articles = [];
+
+        //      // this pushes the returned array to articles
+        //      for (var i = 0; i < data.length; i++) {
+        //          // push data to article
+        //          $scope.articles.push(data[i]);
+
+
+
+        //          // like calculations
+        //          var articleLikers = data[i].likes; // this is an object of users who like the article
+        //          var results = articleLikers.indexOf(userID); // this is etheir -1 or 0
+        //          var specificArticle = data[i];
+        //          if (results >= 0) {
+        //              specificArticle["isLikedByUser"] = true;
+        //          } else {
+        //              specificArticle["isLikedByUser"] = false;
+        //          }
+
+
+        //          // saved for later
+        //          var savedArticleListIds = $scope.savedArticlesIds
+        //          var specificArticleID = data[i]._id; // the id of the article in question
+        //          var savedArticleResults = savedArticleListIds.indexOf(specificArticleID); // need to do a better check here
+        //          if (savedArticleResults >= 0) {
+        //              specificArticle["isSavedByUser"] = true;
+        //          } else {
+        //              specificArticle["isSavedByUser"] = false;
+        //          }
+
+
+
+        //      };
+
+        //      // store last articles date in scope object
+        //      // THIS IS NOT IT THIS IS WHAT NEEDS TO BE DONE
+        //      var last_element = $scope.articles[$scope.articles.length - 1];
+
+
+        //      // $scope.latestDate = 1436318004568;
+        //      $scope.latestDate = last_element.created;
+        //      console.log("The latest date is " + $scope.latestDate);
+
+
+        //      console.log($scope.articles);
+        //  })
+        //  .error(function (data, status, headers, config) {
+        //      console.log("Error when getting home feed")
+        //  });
+    };
+
+    $scope.getMoreHomeFeed = function (minID) {
+        $scope.noMoreItemsAvailable = false;
+
+
+        var userID = window.localStorage.SourceID;
+        var count = 10;
+        var minDate = $scope.latestDate;
+
+        console.log(minID);
+
+        var homeFeedPacket = {
+            userID: userID,
+            count: count,
+            minID: minDate,
+        }
+
+
+        API.getHomeFeed(homeFeedPacket)
+          .success(function (data, user, status, headers, config) {
+
+              // $scope.articles = [];
+
+              // this pushes the returned array to articles
+              for (var i = 0; i < data.length; i++) {
+                  // push data to article
+                  $scope.articles.push(data[i]);
+
+
+                  // like calculations
+                  var articleLikers = data[i].likes; // this is an object of users who like the article
+                  var results = articleLikers.indexOf(userID); // this is etheir -1 or 0
+                  var specificArticle = data[i];
+                  if (results >= 0) {
+                      specificArticle["isLikedByUser"] = true;
+                  } else {
+                      specificArticle["isLikedByUser"] = false;
+                  }
+
+
+                  // saved for later
+                  var savedArticleListIds = $scope.savedArticlesIds
+                  var specificArticleID = data[i]._id; // the id of the article in question
+                  var savedArticleResults = savedArticleListIds.indexOf(specificArticleID); // need to do a better check here
+                  if (savedArticleResults >= 0) {
+                      specificArticle["isSavedByUser"] = true;
+                  } else {
+                      specificArticle["isSavedByUser"] = false;
+                  }
+
+              };
+
+              $scope.$broadcast('scroll.infiniteScrollComplete');
+              console.log($scope.articles);
+          })
+          .error(function (data, status, headers, config) {
+              console.log("Error when getting home feed")
+          });
+    };
+   
+    $scope.deleteSaved = function (articleID) {
+        var userID = window.localStorage.SourceID;
+
+        var savedArticle = {
+            articleID: articleID,
+            userID: userID
+        };
+
+        console.log(savedArticle);
+
+
+        API.deleteSaved(savedArticle)
+            .success(function (article, status, headers, config) {
+                console.log("article delete");
+                $scope.loadMore();
+            })
+            .error(function (article, status, headers, config) {
+                console.log("Something went wrong")
+            });
+
+    };
+
+    $scope.saveForLater = function (articleID) {
+        console.log("trigger");
+
+        var userID = window.localStorage.SourceID;
+
+        var savedArticle = {
+            articleID: articleID,
+            userID: userID
+        };
+
+        API.saveForLater(savedArticle)
+            .success(function (article, user, status, headers, config) {
+                console.log("article save for later");
+                $scope.loadMore();
+            })
+            .error(function (article, status, headers, config) {
+                console.log("Error when saving the article for later")
+            });
+
+    };
+
+    $scope.likeArticle = function (article) {
+
+        var userID = window.localStorage.SourceID;
+        var articleID = article._id;
+        var articleOwner = article._userID;
+        var articleImageUrl = article.imageUrl;
+
+        var likedArticle = {
+            articleID: articleID,
+            imageUrl: articleImageUrl,
+            userID: userID,
+            created: Date.now(),
+            articleOwner: articleOwner
+        };
+
+        console.log(likedArticle);
+
+        API.likeArticle(likedArticle)
+            .success(function (article, user, status, headers, config) {
+                // make button reflect the change
+                console.log("Article Successfully liked");
+                $scope.loadMore();
+            })
+            .error(function (article, status, headers, config) {
+                console.log("Error when liking the article");
+            });
+
+    };
+
+    $scope.unlikeArticle = function (article) {
+
+        var userID = window.localStorage.SourceID;
+        var articleID = article._id;
+
+        var unlikedArticle = {
+            articleID: articleID,
+            userID: userID
+        };
+
+        console.log(unlikedArticle)
+
+
+        API.putLikes(unlikedArticle)
+            .success(function (article, user, status, headers, config) {
+                console.log("Article Successfully unliked")
+                $scope.loadMore();
+            })
+            .error(function (article, status, headers, config) {
+                console.log("Error when liking the article")
+            });
+
+    };
+    
+    $scope.loadMore = function () {
+       
+        var userID = window.localStorage.SourceID;
+        var minDate = $scope.latestDate;
+
+        var homeFeedPacketPaging = {
+            userID: userID,
+            minID: minDate,
+            pageNumber: $scope.pageNumber,
+            itemsPerPage: $scope.itemsPerPage
+        };
+
+        API.getHomeFeedPaging(homeFeedPacketPaging)
+                .success(function (data) {
+                    if (data.length > 0) {
+                        // this pushes the returned array to articles
+                        for (var i = 0; i < data.length; i++) {
+                            // push data to article
+                            $scope.articles.push(data[i]);
+
+                            // like calculations
+                            var articleLikers = data[i].likes; // this is an object of users who like the article
+                            var results = articleLikers.indexOf(userID); // this is etheir -1 or 0
+                            var specificArticle = data[i];
+                            if (results >= 0) {
+                                specificArticle["isLikedByUser"] = true;
+                            } else {
+                                specificArticle["isLikedByUser"] = false;
+                            }
+
+                            // saved for later
+                            var savedArticleListIds = $scope.savedArticlesIds;
+                            var specificArticleID = data[i]._id; // the id of the article in question
+                            var savedArticleResults = savedArticleListIds.indexOf(specificArticleID); // need to do a better check here
+                            if (savedArticleResults >= 0) {
+                                specificArticle["isSavedByUser"] = true;
+                            } else {
+                                specificArticle["isSavedByUser"] = false;
+                            }
+
+                        };
+                    } else {
+                        $scope.noMoreItemsAvailable = true;
+                    }
+
+                    $scope.$broadcast('scroll.infiniteScrollComplete');
+                })
+        .error(function (data, status, headers, config) {
+            console.log("Error when getting home feed");
+        });
+
+        $scope.pageNumber++;
+    };
+
+    //constructor
+    var init = function () {
+        $scope.getInitialHomeFeed();
+
+        $scope.pageNumber = 1;
+        $scope.itemsPerPage = 10;
+        $scope.loadMore();
+    };
+
+    $scope.$on('$ionicView.afterEnter', function () {
+        init();
+    });
+});
+
