+++ sourceTabs/www/templates/home.html
@@ -1,59 +1,47 @@
 <ion-view class="home">
-
-  <!-- Nav Bar -->
-  <ion-header-bar class="nav">
-    <h1 class="title">S O U R C E</h1>
-    <div class="buttons">
-      <i ui-sref="saved" class="icon ion-bookmark menu-icon-right"></i>
-    </div>
-  </ion-header-bar>
-
-
-
-  <ion-content class="has-header" scroll-event-interval="5" ng-init="getInitialHomeFeed()">
-
-    <!-- Refresher -->
-    <ion-refresher
-      pulling-text="Refreshing . . ."
-      on-refresh="doRefresh()">
-    </ion-refresher>
-
-    <ion-list>
-        <div class="list article-card" ng-repeat="articles in articles" >
-          <a href="#/tab/user/{{articles._userID}}">
-              <div class="item item-avatar" style="border-bottom:none">
-                <img src="{{articles.gravatarURL}}">
-                  <div class="float-left username">{{articles.username}}</div>
-                  <p class="float-right date">{{ articles.created | date : 'short' : timezone}}</p>
-              </div>
-          </a>
-
-          <a ui-sref="tabs.readerHome( { articleID: articles._id} )">
-            <img class="full-image" src="{{articles.imageUrl}}">
-          </a>
-
-
-          <div class="article-card-inside">
-              <h2 class="article-title">{{articles.title}}</h2>
-              <p>{{articles.shortUrl}}</p>
-              <p class="article-summary">{{articles.summary}}</p>
-              <button class="button icon-left ion-ios-heart-outline button-clear padding-left-none" ng-click="likeArticle(articles)" ng-show="articles.isLikedByUser == false">Like</button>
-              <span class="like-button"><button class="button icon-left ion-ios-heart button-clear padding-left-none heart-pink" ng-click="unlikeArticle(articles)" ng-show="articles.isLikedByUser == true">Unlike</button></span>
-
-              <button class="button icon-left ion-ios-star-outline button-clear" ng-click="saveForLater(articles._id)" ng-if="articles.isSavedByUser == false">Save For Later</button>
-              <span class="save-for-later-button"><button class="button icon-left ion-ios-star button-clear" ng-click="deleteSaved(articles._id)" ng-if="articles.isSavedByUser == true">Remove</button></span>
-          </div>
+    <!-- Nav Bar -->
+    <ion-header-bar class="nav">
+        <h1 class="title">S O U R C E</h1>
+        <div class="buttons">
+            <i ui-sref="saved" class="icon ion-bookmark menu-icon-right"></i>
         </div>
-    </ion-list>
+    </ion-header-bar>
+    <ion-content class="has-header" scroll-event-interval="5" ng-init="getInitialHomeFeed()">
+        <!-- Refresher -->
+        <!--<ion-refresher pulling-text="Refreshing . . ."
+                       on-refresh="doRefresh()">
+        </ion-refresher>-->
+        <div class="list">
+            <div class="list article-card" ng-repeat="articles in articles">
+                <a href="#/tab/user/{{articles._userID}}">
+                    <div class="item item-avatar" style="border-bottom:none">
+                        <img src="{{articles.gravatarURL}}">
+                        <div class="float-left username">{{articles.username}}</div>
+                        <p class="float-right date">{{ articles.created | date : 'short' : timezone}}</p>
+                    </div>
+                </a>
+                <a ui-sref="tabs.readerHome( { articleID: articles._id} )">
+                    <img class="full-image" src="{{articles.imageUrl}}">
+                </a>
 
+                <div class="article-card-inside">
+                    <h2 class="article-title">{{articles.title}}</h2>
+                    <p>{{articles.shortUrl}}</p>
+                    <p class="article-summary">{{articles.summary}}</p>
+                    <button class="button icon-left ion-ios-heart-outline button-clear padding-left-none" ng-click="likeArticle(articles)" ng-show="articles.isLikedByUser == false">Like</button>
+                    <span class="like-button"><button class="button icon-left ion-ios-heart button-clear padding-left-none heart-pink" ng-click="unlikeArticle(articles)" ng-show="articles.isLikedByUser == true">Unlike</button></span>
+                    <button class="button icon-left ion-ios-star-outline button-clear" ng-click="saveForLater(articles._id)" ng-if="articles.isSavedByUser == false">Save For Later</button>
+                    <span class="save-for-later-button">
+                        <button class="button icon-left ion-ios-star button-clear" ng-click="deleteSaved(articles._id)" ng-if="articles.isSavedByUser == true">Remove</button>
+                    </span>
+                </div>
+            </div>
+        </div>
+        
+        <ion-infinite-scroll ng-if="!noMoreItemsAvailable" on-infinite="loadMore()" distance="10%"></ion-infinite-scroll>
 
-    <ion-infinite-scroll
-      ng-if="!noMoreItemsAvailable"
-      on-infinite="getMoreHomeFeed()"
-      distance="10%">
-    </ion-infinite-scroll>
-
-
-
-  </ion-content>
+        <!--<ion-infinite-scroll on-infinite="getMoreHomeFeed()"
+                             distance="50%">
+        </ion-infinite-scroll>-->
+    </ion-content>
 </ion-view>
